pyproject.toml
[tool.poetry]
name = "fastapi-gemini"
version = "0.2.0"
description = "Poetry venv for Gemini pet project"
authors = ["h1KkkaaH <sav1212121212@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
sqlalchemy = "^2.0.46"
google-genai = "^1.60.0"
python-dotenv = "^1.2.1"
loguru = "^0.7.3"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
httpx = "^0.28.1"
pydantic-settings = "^2.12.0"
passlib = "^1.7.4"
bcrypt = "3.2.2"
asyncpg = "^0.31.0"
email-validator = "^2.2.0"
python-jose = "^3.4.0"
aiosqlite = "^0.20.0"
greenlet = "^3.0.3"
python-multipart = "^0.0.20"
alembic = "^1.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
